Rejigger config so tsx is only in web/
