+++ common/package.json
@@ -12,8 +12,7 @@
     "@tiptap/extension-link": "2.0.0-beta.43",
     "@tiptap/extension-mention": "2.0.0-beta.102",
     "@tiptap/starter-kit": "2.0.0-beta.190",
-    "lodash": "4.17.21",
-    "react-twitter-embed": "4.0.4"
+    "lodash": "4.17.21"
   },
   "devDependencies": {
     "@types/lodash": "4.14.178"
