+++ common/tsconfig.json
@@ -8,8 +8,7 @@
     "outDir": "lib",
     "sourceMap": true,
     "strict": true,
-    "target": "es2017",
-    "jsx": "preserve"
+    "target": "es2017"
   },
-  "include": ["**/*.ts", "util/tweet-embed.tsx"]
+  "include": ["**/*.ts"]
 }
