+++ common/util/parse.ts
@@ -22,7 +22,7 @@ import { Image } from '@tiptap/extension-image'
 import { Link } from '@tiptap/extension-link'
 import { Mention } from '@tiptap/extension-mention'
 import Iframe from './tiptap-iframe'
-import TiptapTweet from './tiptap-tweet'
+import TiptapTweet from './tiptap-tweet-type'
 import { uniq } from 'lodash'
 
 export function parseTags(text: string) {
